setup backend to host static files from CRT build output
