+++ backend/package.json
@@ -4,7 +4,8 @@
   "description": "The backend (and web server site) for the ECC christmas walkthrough hub",
   "main": "index.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc && cp -r ./node_modules/eastgate-cwt-frontend/build ./dist/server/"
   },
   "repository": {
     "type": "git",
@@ -17,6 +18,7 @@
   },
   "homepage": "https://github.com/margeobur/eastgate-cwt#readme",
   "dependencies": {
+    "eastgate-cwt-frontend": "file:../frontend",
     "express": "^4.17.1",
     "ws": "^7.2.0"
   },
